fix(rust): DRY out quaternion/matrix math by re-exporting from math-primitives
