+++ rust_core/math-primitives/src/lib.rs
@@ -13,10 +13,12 @@
 //! - All rotation matrices are SO(3) (orthonormal, det=+1).
 //! - All primitive dimensions must be positive.
 
+pub mod types;
 pub mod geometry;
 #[cfg(feature = "python")]
 pub mod py_bindings;
 pub mod quaternion;
+pub mod matrix3;
 pub mod rotation;
 pub mod transform;
 
@@ -25,9 +27,8 @@ pub use geometry::{
     sphere_capsule_distance, sphere_sphere_distance, Capsule, Cylinder, DistanceResult,
     GeometricPrimitive, OrientedBox, Sphere,
 };
-pub use quaternion::{
-    quaternion_inverse, quaternion_multiply, quaternion_normalize, slerp, Quaternion,
-};
+pub use quaternion::Quaternion;
+pub use matrix3::Matrix3;
 pub use rotation::{
     axis_angle_to_rotation_matrix, euler_to_quaternion, euler_to_rotation_matrix,
     quaternion_to_euler, quaternion_to_rotation_matrix, rotation_matrix_to_euler,
