+++ rust_core/math-primitives/src/matrix3.rs
