+++ rust_core/tools-core/Cargo.toml
@@ -14,14 +14,16 @@ crate-type = ["cdylib", "rlib"]
 [features]
 default = []
 # Enable PyO3 bindings for Python wheel builds (maturin)
-python = ["dep:pyo3"]
+# Enable PyO3 bindings for Python wheel builds (maturin)
+python = ["dep:pyo3", "math-primitives/python"]
 # Enable wasm-bindgen for WASM/NPM builds (wasm-pack)
-wasm = ["dep:wasm-bindgen"]
+wasm = ["dep:wasm-bindgen", "math-primitives/wasm"]
 
 [dependencies]
 pyo3 = { workspace = true, optional = true }
 wasm-bindgen = { workspace = true, optional = true }
 serde = { workspace = true }
+math-primitives = { path = "../math-primitives" }
 
 [dev-dependencies]
 proptest = { workspace = true }
