+++ rust_core/tools-core/src/ball_flight.rs
@@ -24,7 +24,7 @@
 
 use serde::{Deserialize, Serialize};
 
-use crate::types::Vector3;
+use math_primitives::types::Vector3;
 
 // ── Physical Constants ───────────────────────────────────────────────────────
