+++ rust_core/tools-core/src/lib.rs
@@ -3,43 +3,25 @@
 //! This crate provides the canonical implementations of mathematical primitives,
 //! physics solvers, and thermodynamic calculators used across all AffineDrift
 //! repositories. It compiles to:
-//!
-//! - A Python extension module (via PyO3 / Maturin)
-//! - A WebAssembly module (via wasm-bindgen / wasm-pack)
-//! - A native Rust library (for direct Rust consumers)
-//!
-//! ## Design Principles
-//!
-//! - **TDD**: Every public function has tests written before implementation.
-//! - **DbC (Design by Contract)**: All public functions validate preconditions
-//!   via `debug_assert!` and return `Result<T, E>` instead of panicking.
-//! - **DRY**: Types and logic defined once here; Python/JS are thin wrappers.
 
 pub mod atmosphere;
 pub mod ball_flight;
 pub mod math;
-pub mod matrix3;
-pub mod quaternion;
-pub mod types;
-
-// Re-export primary types at crate root for ergonomic imports.
+// Re-export primary types
 pub use math::{clamp, lerp, GRAVITY, R_GAS};
-pub use matrix3::Matrix3;
-pub use quaternion::Quaternion;
-pub use types::Vector3;
+pub use math_primitives::matrix3::Matrix3;
+pub use math_primitives::quaternion::Quaternion;
+pub use math_primitives::types::Vector3;
 
-// ── Python bindings (feature-gated) ──────────────────────────────────────────
-
-/// Register the Python module when compiled with `--features python`.
 #[cfg(feature = "python")]
 use pyo3::prelude::*;
 
 #[cfg(feature = "python")]
 #[pymodule]
 fn tools_core(m: &Bound<'_, PyModule>) -> PyResult<()> {
-    m.add_class::<types::Vector3>()?;
-    m.add_class::<quaternion::Quaternion>()?;
-    m.add_class::<matrix3::Matrix3>()?;
+    m.add_class::<math_primitives::types::Vector3>()?;
+    m.add_class::<math_primitives::quaternion::Quaternion>()?;
+    m.add_class::<math_primitives::matrix3::Matrix3>()?;
     m.add_class::<ball_flight::BallProperties>()?;
     m.add_class::<ball_flight::LaunchConditions>()?;
     m.add_class::<ball_flight::EnvironmentalConditions>()?;
